refact: DRY agent-test -> TestSupport
